rag-production-improved/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
transformers
torch
numpy
pytest
sentencepiece
accelerate
nltk
langfuse==2.60.0



